feat: navigation 컴포넌트 최종 구현(슬라이드 애니메이션 제외)
