+++ packages/design-system/src/components/navigation/navigation.css.ts
@@ -42,6 +42,11 @@ export const list = recipe({
   },
 });
 
-export const panel = style({
-  display: 'flex',
+export const underBar = style({
+  width: '100%',
+  height: '2px',
+  backgroundColor: themeVars.color.confeti_lime,
+  position: 'absolute',
+  left: 0,
+  bottom: 0,
 });
