--- packages/design-system/src/components/navigation/under-bar/under-bar.tsx
@@ -1,7 +1,0 @@
-import * as styles from './under-bar.css';
-
-const UnderBar = () => {
-  return <div className={styles.underBar}></div>;
-};
-
-export default UnderBar;
